Moved security classes to config package
